try hotplug by default
This commit enables hotplug by default for 'gnt-instance modify' and adds the option --no-hotplug to disable hotplug.
A restart should not always be carried out when parameters are changed if this is not necessary.

The parameter --hotplug_if_possible has been removed. Hotplug is now always tried with gnt-instance modify.
A warning is now displayed instead of an error.

See #1548

Signed-off-by: codefritzel <pascal-pf@hotmail.de>